Scripts/Foundry: Add first work on Iron Assembly + Beastlord Darmac

+++ src/server/game/Achievements/AchievementMgr.h
@@ -546,7 +546,7 @@ struct AchievementCriteriaData
     }
 
     bool IsValid(CriteriaEntry const* criteria);
-    bool Meets(uint32 criteria_id, Player const* source, Unit const* target, uint32 miscvalue1 = 0) const;
+    bool Meets(uint32 criteria_id, Player const* source, Unit const* target, uint64 miscvalue1 = 0) const;
 };
 
 struct AchievementCriteriaDataSet
@@ -554,7 +554,7 @@ struct AchievementCriteriaDataSet
         AchievementCriteriaDataSet() : criteria_id(0) {}
         typedef std::vector<AchievementCriteriaData> Storage;
         void Add(AchievementCriteriaData const& data) { storage.push_back(data); }
-        bool Meets(Player const* source, Unit const* target, uint32 miscvalue = 0) const;
+        bool Meets(Player const* source, Unit const* target, uint64 miscvalue = 0) const;
         void SetCriteriaId(uint32 id) {criteria_id = id;}
     private:
         uint32 criteria_id;
